chore: integrate babel
