+++ src/cjs.js
@@ -1,0 +1,3 @@
+const loader = require('./index');
+
+module.exports = loader.default;
