+++ src/index.js
@@ -1,0 +1,16 @@
+import { getOptions } from 'loader-utils';
+import validateOptions from 'schema-utils';
+
+import schema from './options.json';
+
+export default function rawLoader(source) {
+  const options = getOptions(this) || {};
+
+  validateOptions(schema, options, 'Raw Loader');
+
+  const json = JSON.stringify(source)
+    .replace(/\u2028/g, '\\u2028')
+    .replace(/\u2029/g, '\\u2029');
+
+  return `module.exports = ${json}`;
+}
